+++ lib/loading_animation_widget.dart
@@ -11,16 +11,16 @@ import 'src/hexagon_dots/hexagon_dots.dart';
 import 'src/flickr/flickr.dart';
 import 'src/half_tringle_dot/half_tringle_dot.dart';
 
-import 'src/horizontalDotsRotation/horizontal_dots_rotation.dart';
+import 'src/horizontal_rotating_dots/horizontal_rotating_dots.dart';
 import 'src/ink_drop/ink_drop.dart';
 import 'src/newton_cradle/newton_cradle.dart';
 import 'src/twisting_dots/twisting_dots.dart';
-import 'src/staggered_wave/staggered_wave.dart';
+import 'src/staggered_dots_wave/staggered_dots_wave.dart';
 import 'src/stretched_dots/stretched_dots.dart';
-import 'src/waving_dots/waving_dots.dart';
+import 'src/wave_dots/wave_dots.dart';
 
 import 'src/bouncing_ball/bouncing_ball.dart';
-import 'src/arch_circle/arch_circle.dart';
+import 'src/three_arched_circle/three_arched_circle.dart';
 import 'src/prograssive_dots/prograssive_dots.dart';
 import 'src/four_rotating_dots/four_rotating_dots.dart';
 import 'src/falling_dot/falling_dot.dart';
@@ -29,296 +29,295 @@ class LoadingAnimationWidget {
   LoadingAnimationWidget._();
 
   /// A ball falls inside a ring cut out when face up and falls when face down.
-  /// Required color is applied to the ring and dot. Time is in milliseconds which is optional.
+  /// Required color is applied to the ring and dot.
   static Widget fallingDot({
     required Color color,
     required double size,
-    int time = 1500,
+    Key? key,
   }) {
     return FallingDot(
       color: color,
       size: size,
-      time: time,
+      key: key,
     );
   }
 
   /// Four dots shrink to the center and expand and rotates 315 degree.
-  /// Required color is applied to four dots. Time is in milliseconds which is optional.
+  /// Required color is applied to four dots.
   static Widget fourRotatingDots({
     required Color color,
     required double size,
-    int time = 2200,
+    Key? key,
   }) {
     return FourRotatingDots(
       color: color,
       size: size,
-      time: time,
+      key: key,
     );
   }
 
   /// Four dots in a row left most scales down and a new dot emerge from right.
-  /// Required color is applied to four dots. Time is in milliseconds which is optional.
+  /// Required color is applied to four dots.
   static Widget prograssiveDots({
     required Color color,
     required double size,
-    int time = 800,
+    Key? key,
   }) {
     return PrograssiveDots(
       color: color,
       size: size,
-      time: time,
+      key: key,
     );
   }
 
   /// A dot becomes a full ring and also have discrete following rings.
-  /// Required color is applied to the circular ring. Time is in milliseconds which is optional.
+  /// Required color is applied to the circular ring.
   static Widget discreteCircle({
     required Color color,
     required double size,
-    int time = 2000,
+    Color secondRingColor = Colors.teal,
+    Color thirdRingColor = Colors.orange,
+    Key? key,
   }) {
     return DiscreteCircle(
       color: color,
       size: size,
-      time: time,
+      secondCircleColor: secondRingColor,
+      thirdCircleColor: thirdRingColor,
+      key: key,
     );
   }
 
   /// Three arc at 60 degree shrinks to a dot while rotating and comes back to inital.
-  /// Required color is applied to all three arc. Time is in milliseconds which is optional.
+  /// Required color is applied to all three arc.
   static Widget threeArchedCircle({
     required Color color,
     required double size,
-    int time = 2000,
+    Key? key,
   }) {
     return ThreeArchedCircle(
       color: color,
       size: size,
-      time: time,
+      key: key,
     );
   }
 
   /// A ball fall from top and bounce back
-  /// Required color is applied to the ball. Time is in milliseconds which is optional.
-
+  /// Required color is applied to the ball.
   static Widget bouncingBall({
     required Color color,
     required double size,
-    int time = 1200,
+    Key? key,
   }) {
     return BouncingBall(
       color: color,
       size: size,
-      time: time,
+      key: key,
     );
   }
 
   /// Similar to flickr loading animation. Two disks swap positon and come back to initial.
   /// Required leftDotColor is applied to the left disk.
   /// Required rightDotColor is applied to the right disk.
-  /// Time is in milliseconds which is optional.
   static Widget flickr({
     required Color leftDotColor,
     required Color rightDotColor,
     required double size,
-    int time = 2000,
+    Key? key,
   }) {
     return Flickr(
       leftDotColor: leftDotColor,
       rightDotColor: rightDotColor,
       size: size,
-      time: time,
+      key: key,
     );
   }
 
   /// Six dots in each vertice of a hexagon emerge one by one then shrink to disappear while rotating
-  /// Required color is applied to all dots. Time is in milliseconds which is optional.
+  /// Required color is applied to all dots.
   static Widget hexagonDots({
     required Color color,
     required double size,
-    int time = 3000,
+    Key? key,
   }) {
     return HexagonDots(
       color: color,
       size: size,
-      time: time,
+      key: key,
     );
   }
 
   /// One ring emerge from the center and scale up until touches the outer ring.
   /// Then the outer ring expand a bit then come back to normal.
-  /// Required color is applied to both rings. Time is in milliseconds which is optional.
+  /// Required color is applied to both rings.
   static Widget beat({
     required Color color,
     required double size,
-    int time = 1000,
+    Key? key,
   }) {
     return Beat(
       color: color,
       size: size,
-      time: time,
+      key: key,
     );
   }
 
   /// Three dots in each vertice of an equilateral triangle expand towards side
   /// then moves to next vertice. Required color is applied to all dots.
-  /// Time is in milliseconds which is optional.
   static Widget dotsTriangle({
     required Color color,
     required double size,
-    int time = 1200,
+    Key? key,
   }) {
     return DotsTriangle(
       color: color,
       size: size,
-      time: time,
+      key: key,
     );
   }
 
   /// An equilateral triangle has two side and a dot in the empty side. The dot
   /// moves those two sides also change position. Required color is applied to
-  /// two sides and the dot. Time is in milliseconds which is optional.
+  /// two sides and the dot.
   static Widget halfTringleDot({
     required Color color,
     required double size,
-    int time = 3000,
+    Key? key,
   }) {
     return HalfTringleDot(
       color: color,
       size: size,
-      time: time,
+      key: key,
     );
   }
 
   /// Two arc place at 180 degree apart from each other. While rotating shrink to dot.
-  /// Required color is applied to both arc. Time is in milliseconds which is optional.
+  /// Required color is applied to both arc.
   static Widget twoRotatingArc({
     required Color color,
     required double size,
-    int time = 1500,
+    Key? key,
   }) {
     return TwoRotatingArc(
       color: color,
       size: size,
-      time: time,
+      key: key,
     );
   }
 
   /// Three dots rotates in a horizontal plane. Required color is applied to all
-  /// three dots. Time is in milliseconds which is optional.
-  static Widget threeHorizontalDots({
+  /// three dots.
+  static Widget horizontalRotatingDots({
     required Color color,
     required double size,
-    int time = 800,
+    Key? key,
   }) {
-    return ThreeHorizontalDots(
+    return HorizontalRotatingDots(
       color: color,
       size: size,
-      time: time,
+      key: key,
     );
   }
 
   /// Four balls swing and give newton cradle animation. Required color is
-  /// applied to all four balls. Time is in milliseconds which is optional.
+  /// applied to all four balls.
   static Widget newtonCradle({
     required Color color,
     required double size,
-    int time = 1200,
+    Key? key,
   }) {
     return NewtonCradle(
       color: color,
       size: size,
-      time: time,
+      key: key,
     );
   }
 
   /// A dot falls down then completes a circle then become to dot again.
-  /// Required color is applied to the ring and the dot. Time is in
-  /// milliseconds which is optional.
+  /// Required color is applied to the ring and the dot.
   static Widget inkDrop({
     required Color color,
     required double size,
-    int time = 1400,
+    Key? key,
   }) {
     return InkDrop(
       color: color,
       size: size,
-      time: time,
+      key: key,
     );
   }
 
   /// Two disks rotate twisting. Required leftDotColor is applied to the left
   /// disk. Required leftDotColor is applied to the left disk. Required rightDotColor
-  /// is applied to the right disk. Time is in milliseconds which is optional.
+  /// is applied to the right disk.
   static Widget twistingDots({
     required Color leftDotColor,
     required Color rightDotColor,
     required double size,
-    int time = 3000,
+    Key? key,
   }) {
     return TwistingDots(
       leftDotColor: leftDotColor,
       rightDotColor: rightDotColor,
       size: size,
-      time: time,
+      key: key,
     );
   }
 
   /// Five dots expands one after another like wave and then come back to
-  /// original. Required color is applied to all dots. Time is in
-  /// milliseconds which is optional.
-  static Widget staggeredDotWave({
+  /// original. Required color is applied to all dots.
+  static Widget staggeredDotsWave({
     required Color color,
     required double size,
-    int time = 1600,
+    Key? key,
   }) {
-    return StaggeredDotWave(
+    return StaggeredDotsWave(
       color: color,
       size: size,
-      time: time,
+      key: key,
     );
   }
 
   /// Five dots stretch up and shrink. Then again stretch down and shrink. Required
-  /// color is applied to all dots. Time is in milliseconds which is optional.
+  /// color is applied to all dots.
   static Widget stretchedDots({
     required Color color,
     required double size,
-    int time = 2000,
+    Key? key,
   }) {
     return StretchedDots(
       color: color,
       size: size,
-      time: time,
+      key: key,
     );
   }
 
   /// Three dots go up and down give wave effect. Required color is applied to
-  /// all dots. Time is in milliseconds which is optional.
-  static Widget wavingDots({
+  /// all dots.
+  static Widget waveDots({
     required Color color,
     required double size,
-    int time = 1200,
+    Key? key,
   }) {
-    return WavingDots(
+    return WaveDots(
       color: color,
       size: size,
-      time: time,
+      key: key,
     );
   }
 
   /// Three balls at each vertice of an equilateral triangle moves to the center
   /// while rotating and then again go back to origianl position. Required color
-  /// is applied to all balls. Time is in milliseconds which is optional.
-  static Widget threeRotatingBalls({
+  /// is applied to all balls.
+  static Widget threeRotatingDots({
     required Color color,
     required double size,
-    int time = 2000,
+    Key? key,
   }) {
-    return ThreeRotatingBalls(
+    return ThreeRotatingDots(
       color: color,
       size: size,
-      time: time,
+      key: key,
     );
   }
 }