+++ lib/src/discrete_circle/discrete_circle.dart
@@ -5,12 +5,14 @@ import 'dart:math' as math;
 class DiscreteCircle extends StatefulWidget {
   final double size;
   final Color color;
-  final int time;
+  final Color secondCircleColor;
+  final Color thirdCircleColor;
   const DiscreteCircle({
     Key? key,
     required this.color,
     required this.size,
-    required this.time,
+    required this.secondCircleColor,
+    required this.thirdCircleColor,
   }) : super(key: key);
 
   @override
@@ -26,7 +28,7 @@ class _DiscreteCircleState extends State<DiscreteCircle>
     super.initState();
     _animationController = AnimationController(
       vsync: this,
-      duration: Duration(milliseconds: widget.time),
+      duration: const Duration(milliseconds: 2000),
     )..repeat();
   }
 
@@ -35,6 +37,8 @@ class _DiscreteCircleState extends State<DiscreteCircle>
     final Color color = widget.color;
     final double size = widget.size;
     final double strokeWidth = size / 8;
+    final Color secondRingColor = widget.secondCircleColor;
+    final Color thirdRingColor = widget.thirdCircleColor;
     return AnimatedBuilder(
       animation: _animationController,
       builder: (_, __) {
@@ -56,7 +60,7 @@ class _DiscreteCircleState extends State<DiscreteCircle>
               child: Visibility(
                 visible: _animationController.value >= 0.5,
                 child: Arc.draw(
-                  color: Colors.orange,
+                  color: thirdRingColor,
                   size: size,
                   strokeWidth: strokeWidth,
                   startAngle: -math.pi / 2,
@@ -81,7 +85,7 @@ class _DiscreteCircleState extends State<DiscreteCircle>
             Visibility(
               visible: _animationController.value >= 0.5,
               child: Arc.draw(
-                color: Colors.teal,
+                color: secondRingColor,
                 size: size,
                 strokeWidth: strokeWidth,
                 startAngle: -math.pi / 2,