+++ test/base_structure.dart
@@ -1,0 +1,10 @@
+import 'package:flutter/material.dart';
+
+
+
+Widget baseStructure(Widget animationWidget) => MaterialApp(
+      home: Material(
+        child: animationWidget,
+      ),
+    );
+